platform/ops/scripts/orchestration/requirements-airflow.txt
# Core dependencies with pinned versions to avoid conflicts
loguru==0.7.2
mlflow==2.17.2
pandas==2.2.3
scikit-learn==1.5.2
xgboost==2.1.3
boto3==1.35.82

# ClickHouse
clickhouse-connect==0.8.18
clickhouse-driver==0.2.9

# Fix grpcio conflict (pin compatible versions)
grpcio==1.62.2
grpcio-status==1.62.2
protobuf==4.25.3

# Airflow providers (pin to avoid backtracking)
apache-airflow-providers-apache-spark==4.10.0
apache-airflow-providers-openlineage==1.11.0
apache-airflow-providers-docker==3.14.0
apache-airflow-providers-amazon

# dbt
dbt-core==1.8.7
dbt-clickhouse==1.8.2